requirements.txt
torch~=2.1.1
pytorch-lightning~=2.1.1
transformers==4.45.2
numpy==1.24.4
pandas==2.0.3
matplotlib==3.7.1
fastapi==0.109.1
uvicorn==0.15.0
requests==2.31.0
setuptools>=70.0.0
scikit-learn==1.3.0
seaborn==0.11.2
pydantic==1.10.18
python-jose==3.3.0
python-dotenv==1.0.0
pytest==7.4.3
rich==13.9.0
click==8.1.7
optuna~=4.1.0
config~=0.5.1
PyJWT~=2.10.0
prometheus-client~=0.21.0
ploty~=5.24.1
scipy~=1.13.1
neptune~=2.0.0a9
